role包添加，添加角色功能，但还未整合

+++ src/main/java/com/aiwan/util/SMToDecodeData.java
@@ -6,7 +6,7 @@ import com.aiwan.publicsystem.protocol.DecodeData;
  * DecodeData类转换工具类
  * */
 public class SMToDecodeData {
-    public static DecodeData shift(short type,Object object){
+    public static DecodeData shift(int type,Object object){
         DecodeData decodeData = new DecodeData();
         decodeData.setType(type);
         byte[] data = ObjectToBytes.writeInto(object);
